refactor: 로또 수 검증의 책임은 LottoNumber로 이동
