+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -2,12 +2,10 @@ package lotto.controller
 
 import lotto.domain.Amount
 import lotto.domain.LottoNumber
-import lotto.domain.LottoNumberGenerator
 import lotto.domain.LottoResult
 import lotto.domain.LottoShop
 import lotto.domain.LottoTicket
 import lotto.domain.WinningLotto
-import lotto.error.CustomException
 
 class LottoController(
     private val shop: LottoShop = LottoShop(),
@@ -22,7 +20,7 @@ class LottoController(
         val purchasedTicket =
             purchasedTicket ?: return EndLottoResponse(LottoResult.withoutPurchasedTicket())
         val winningLotto = WinningLotto(
-            winningNumber = LottoNumberGenerator.createFrom(request.winningNumbers),
+            winningNumber = LottoNumber.of(request.winningNumbers),
             bonusNumber = request.bonusNumber
         )
         return shop.receivePrize(purchasedTicket, winningLotto).let(::EndLottoResponse)
@@ -31,5 +29,5 @@ class LottoController(
     private fun LottoTicket.save(): LottoTicket =
         this.also { purchasedTicket = it }
 
-    private fun List<List<Int>>.toLottoNumbers(): List<LottoNumber> = this.map { LottoNumberGenerator.createFrom(it) }
+    private fun List<List<Int>>.toLottoNumbers(): List<LottoNumber> = this.map { LottoNumber.of(it) }
 }
